chore: put privateKeyPackedData into JsonWebKey
