+++ Maskbook/Persona/JsonWebKey.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import SwiftMsgPack
 
 struct JsonWebKey: Codable {
     struct RsaOtherPrimesInfo: Codable {
@@ -43,5 +44,38 @@ extension JsonWebKey {
         } else {
             return nil
         }
+    }
+}
+
+extension JsonWebKey {
+    var privateKeyPackedData: Data? {
+        var data = Data()
+        guard let key_ops = key_ops else { return nil }
+        do {
+            let type = MsgPackType.dict(items: 7)
+            var type_value = try type.value()
+            data.append( UnsafeBufferPointer(start: &type_value, count: 1) )
+            _ = try data.pack("y")
+            _ = try data.pack(y ?? "")
+            _ = try data.pack("x")
+            _ = try data.pack(x ?? "")
+            _ = try data.pack("key_ops")
+            _ = try data.pack(key_ops)
+            _ = try data.pack("kty")
+            _ = try data.pack(kty ?? "")
+            _ = try data.pack("ext")
+            _ = try data.pack(true)
+            _ = try data.pack("crv")
+            _ = try data.pack(crv ?? "")
+            _ = try data.pack("d")
+            _ = try data.pack(d ?? "")
+        } catch {
+            print(error)
+        }
+        return data
+    }
+    
+    var privateKeyBase64String: String? {
+        return privateKeyPackedData?.base64EncodedString()
     }
 }
